third_party/blink/web_tests/platform/linux/virtual/dom-parts-disabled/external/wpt/dom/parts/basic-dom-part-declarative-pi-syntax.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Basic declarative DOM Parts (Main Document)
  doc.getPartRoot is not a function
[FAIL] Basic declarative DOM Parts (Template)
  doc.getPartRoot is not a function
[FAIL] Basic declarative DOM Parts (Clone)
  doc.getPartRoot is not a function
[FAIL] Basic declarative DOM Parts (PartClone)
  template.content.getPartRoot is not a function
Harness: the test ran to completion.
